Move color legend to be above both visualisations
