+++ ui/packages/shared/profile/src/ProfileIcicleGraph/index.tsx
@@ -13,20 +13,18 @@
 
 import React, {useEffect, useMemo, useState} from 'react';
 
-import {Table, tableFromIPC} from 'apache-arrow';
 import {AnimatePresence, motion} from 'framer-motion';
 
 import {Flamegraph, FlamegraphArrow} from '@parca/client';
 import {IcicleGraphSkeleton, useParcaContext, useURLState} from '@parca/components';
 import {ProfileType} from '@parca/parser';
-import {capitalizeOnlyFirstLetter, divide, selectQueryParam} from '@parca/utilities';
+import {capitalizeOnlyFirstLetter, divide} from '@parca/utilities';
 
 import {useProfileViewContext} from '../ProfileView/ProfileViewContext';
 import DiffLegend from '../components/DiffLegend';
 import {IcicleGraph} from './IcicleGraph';
 import {FIELD_FUNCTION_NAME, IcicleGraphArrow} from './IcicleGraphArrow';
-import ColorStackLegend from './IcicleGraphArrow/ColorStackLegend';
-import useMappingList, {useFilenamesList} from './IcicleGraphArrow/useMappingList';
+import useMappingList from './IcicleGraphArrow/useMappingList';
 
 const numberFormatter = new Intl.NumberFormat('en-US');
 
@@ -70,14 +68,8 @@ const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
   const {onError, authenticationErrorMessage, isDarkMode} = useParcaContext();
   const {compareMode} = useProfileViewContext();
   const [isLoading, setIsLoading] = useState<boolean>(true);
-  const isColorStackLegendEnabled = selectQueryParam('color_stack_legend') === 'true';
-
-  const table: Table<any> | null = useMemo(() => {
-    return arrow !== undefined ? tableFromIPC(arrow.record) : null;
-  }, [arrow]);
 
   const mappingsList = useMappingList(metadataMappingFiles);
-  const filenamesList = useFilenamesList(table);
 
   const [storeSortBy = FIELD_FUNCTION_NAME] = useURLState('sort_by');
   const [colorBy, setColorBy] = useURLState('color_by');
@@ -89,7 +81,6 @@ const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
   const [compareAbsolute = compareAbsoluteDefault] = useURLState('compare_absolute');
   const isCompareAbsolute = compareAbsolute === 'true';
 
-  const colorByValue = colorBy === undefined || colorBy === '' ? 'binary' : (colorBy as string);
   const mappingsListCount = useMemo(
     () => mappingsList.filter(m => m !== '').length,
     [mappingsList]
@@ -231,13 +222,6 @@ const ProfileIcicleGraph = function ProfileIcicleGraphNonMemo({
         transition={{duration: 0.5}}
       >
         {compareMode ? <DiffLegend /> : null}
-        {isColorStackLegendEnabled && (
-          <ColorStackLegend
-            compareMode={compareMode}
-            mappings={colorByValue === 'binary' ? mappingsList : filenamesList}
-            loading={isLoading}
-          />
-        )}
         <div className="min-h-48" id="h-icicle-graph">
           <>{icicleGraph}</>
         </div>
